DLSV2-104: Moves _layout.cshtml to Views/Shared folder and repurposes for use across all DLS apps. Menu items for a specific app are brought in as a partial view in the Shared folder for that app (e.g. _NavMenuItems.cshtml). Error pages are also moved to Views/Shared/Error and repurposed to make generic (links are to App selector rather than LearningPortal/Current) for example.
