+++ DigitalLearningSolutions.Web/Startup.cs
@@ -93,8 +93,8 @@ namespace DigitalLearningSolutions.Web
                 app.UseBrowserLink();
             }
 
-            app.UseExceptionHandler("/LearningPortal/Error");
-            app.UseStatusCodePagesWithReExecute("/LearningPortal/StatusCode/{0}");
+            app.UseExceptionHandler("/Frameworks/Error");
+            app.UseStatusCodePagesWithReExecute("/Frameworks/StatusCode/{0}");
             app.UseStaticFiles();
             app.UseSerilogRequestLogging();
             app.UseRouting();
