--- DigitalLearningSolutions.Web/ViewModels/LearningPortal/HeaderViewModel.cs
@@ -1,20 +1,0 @@
-﻿namespace DigitalLearningSolutions.Web.ViewModels.LearningPortal
-{
-    using DigitalLearningSolutions.Data.Models;
-
-    public class HeaderViewModel
-    {
-        public readonly string? CustomLogoDataUrl;
-        public readonly int Height;
-        public readonly int Width;
-        public readonly string AltText;
-
-        public HeaderViewModel(CentreLogo? customLogo)
-        {
-            CustomLogoDataUrl = customLogo?.LogoUrl;
-            Height = customLogo?.Height ?? 0 ;
-            Width = customLogo?.Width ?? 0;
-            AltText = customLogo?.CentreName ?? "";
-        }
-    }
-}
