+++ DigitalLearningSolutions.Web/Views/LearningPortal/Available/Available.cshtml
@@ -4,9 +4,12 @@
 <link rel="stylesheet" href="@Url.Content("~/css/learningPortal/coursesShared.css")" asp-append-version="true">
 
 @{
-  ViewData["Title"] = "Learning Portal - Available";
+    ViewData["Application"] = "Learning Portal";
+    ViewData["Title"] = "Learning Portal - Available";
 }
-
+@section NavMenuItems {
+  <partial name="Shared/_NavMenuItems" />
+}
 <div class="nhsuk-grid-row">
   <div class="nhsuk-grid-column-full">
     <h1 id="page-heading">Available Activities</h1>
@@ -25,7 +28,7 @@
       }
       else
       {
-        <partial name="Shared/_SearchResultsCount" model="Model"/>
+        <partial name="Shared/_SearchResultsCount" model="Model" />
 
         <div id="course-cards">
           @foreach (var availableCourse in Model.AvailableCourses)
