+++ DigitalLearningSolutions.Web/Views/LearningPortal/Current/Current.cshtml
@@ -6,9 +6,12 @@
 <link rel="stylesheet" href="@Url.Content("~/css/learningPortal/current.css")" asp-append-version="true">
 
 @{
-  ViewData["Title"] = "Learning Portal - Current";
+    ViewData["Application"] = "Learning Portal";
+    ViewData["Title"] = "Learning Portal - Current";
 }
-
+@section NavMenuItems {
+  <partial name="Shared/_NavMenuItems"/>
+}
 <div class="nhsuk-grid-row">
   <div class="nhsuk-grid-column-full">
     <h1 id="page-heading">My Current Activities</h1>
