--- DigitalLearningSolutions.Web/Views/LearningPortal/Error/Forbidden.cshtml
@@ -1,14 +1,0 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.LearningPortal
-@model ErrorViewModel
-@{
-  ViewData["Title"] = "Learning Portal - You don't have permission to view this page";
-}
-
-<div class="nhsuk-grid-row">
-  <div class="nhsuk-grid-column-two-thirds nhsuk-u-reading-width">
-    <h1>You don't have permission to view this page</h1>
-    <p>You do not have the permissions required to view this page.</p>
-    <p>View your <a href="/">current courses</a> or <a href="https://www.dls.nhs.uk/Home">visit the Digital Learning Solutions homepage</a>.</p>
-    <p>@Model.HelpText()</p>
-  </div>
-</div>
