--- DigitalLearningSolutions.Web/Views/LearningPortal/Error/UnknownError.cshtml
@@ -1,14 +1,0 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.LearningPortal
-@model ErrorViewModel
-@{
-  ViewData["Title"] = "Learning Portal - Something went wrong";
-}
-
-<div class="nhsuk-grid-row">
-  <div class="nhsuk-grid-column-two-thirds nhsuk-u-reading-width">
-    <h1>Something went wrong</h1>
-    <p>Please refresh the page to try again. You may need to try again later.</p>
-    <p>You can also <a href="/">view your current courses</a> or <a href="https://www.dls.nhs.uk/Home">visit the Digital Learning Solutions homepage</a>.</p>
-    <p>@Model.HelpText()</p>
-  </div>
-</div>
