+++ DigitalLearningSolutions.Web/Views/LearningPortal/Shared/_NavMenuItems.cshtml
@@ -1,0 +1,25 @@
+﻿@using DigitalLearningSolutions.Web.Extensions
+<li class="nhsuk-header__navigation-item @Html.IsSelected("Current")">
+  <a class="nhsuk-header__navigation-link" asp-controller="LearningPortal" asp-action="Current">
+    Current activities
+    <svg class="nhsuk-icon nhsuk-icon__chevron-right" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" aria-hidden="true">
+      <path d="M15.5 12a1 1 0 0 1-.29.71l-5 5a1 1 0 0 1-1.42-1.42l4.3-4.29-4.3-4.29a1 1 0 0 1 1.42-1.42l5 5a1 1 0 0 1 .29.71z"></path>
+    </svg>
+  </a>
+</li>
+<li class="nhsuk-header__navigation-item @Html.IsSelected("Completed")">
+  <a class="nhsuk-header__navigation-link" asp-controller="LearningPortal" asp-action="Completed">
+    Completed activities
+    <svg class="nhsuk-icon nhsuk-icon__chevron-right" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" aria-hidden="true">
+      <path d="M15.5 12a1 1 0 0 1-.29.71l-5 5a1 1 0 0 1-1.42-1.42l4.3-4.29-4.3-4.29a1 1 0 0 1 1.42-1.42l5 5a1 1 0 0 1 .29.71z"></path>
+    </svg>
+  </a>
+</li>
+<li class="nhsuk-header__navigation-item @Html.IsSelected("Available")">
+  <a class="nhsuk-header__navigation-link" asp-controller="LearningPortal" asp-action="Available">
+    Available activities
+    <svg class="nhsuk-icon nhsuk-icon__chevron-right" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" aria-hidden="true">
+      <path d="M15.5 12a1 1 0 0 1-.29.71l-5 5a1 1 0 0 1-1.42-1.42l4.3-4.29-4.3-4.29a1 1 0 0 1 1.42-1.42l5 5a1 1 0 0 1 .29.71z"></path>
+    </svg>
+  </a>
+</li>
