+++ DigitalLearningSolutions.Web/Views/Shared/Error/Forbidden.cshtml
@@ -1,0 +1,14 @@
+﻿@using DigitalLearningSolutions.Web.ViewModels.LearningPortal
+@model ErrorViewModel
+@{
+  ViewData["Title"] = "Learning Portal - You don't have permission to view this page";
+}
+
+<div class="nhsuk-grid-row">
+  <div class="nhsuk-grid-column-two-thirds nhsuk-u-reading-width">
+    <h1>You don't have permission to view this page</h1>
+    <p>You do not have the permissions required to view this page.</p>
+    <p>Return to the <a href="https://www.dls.nhs.uk/Home">Digital Learning Solutions homepage</a>.</p>
+    <p>@Model.HelpText()</p>
+  </div>
+</div>
