+++ DigitalLearningSolutions.Web/Views/Shared/Error/PageNotFound.cshtml
@@ -1,0 +1,14 @@
+﻿@using DigitalLearningSolutions.Web.ViewModels.LearningPortal
+@model ErrorViewModel
+@{
+  ViewData["Title"] = "Learning Portal - Page not found";
+}
+
+<div class="nhsuk-grid-row">
+  <div class="nhsuk-grid-column-two-thirds nhsuk-u-reading-width">
+    <h1>We can’t find the page you’re looking for</h1>
+    <p>Please check the web address to make sure it was correct.</p>
+    <p>Or return to the <a href="https://www.dls.nhs.uk/Home">Digital Learning Solutions homepage</a>.</p>
+    <p>@Model.HelpText()</p>
+    </div>
+</div>
